refactor: Rename RichTextEditor
