+++ src/script/components/RichTextEditor/components/SendMessageButton/SendMessageButton.tsx
