+++ src/script/components/RichTextEditor/components/SendMessageButton/index.ts
