+++ src/script/components/RichTextEditor/nodes/EmojiNode.ts
