+++ src/script/components/RichTextEditor/nodes/MentionNode.tsx
