+++ src/script/components/RichTextEditor/plugins/AutoFocusPlugin.tsx
