+++ src/script/components/RichTextEditor/plugins/DraftStatePlugin.tsx
