+++ src/script/components/RichTextEditor/plugins/EditMessagePlugin.tsx
