+++ src/script/components/RichTextEditor/plugins/EmojiPickerPlugin/EmojiItem.styles.ts
