+++ src/script/components/RichTextEditor/plugins/EmojiPickerPlugin/emojiList.ts
