+++ src/script/components/RichTextEditor/plugins/EmojiPickerPlugin/index.ts
