+++ src/script/components/RichTextEditor/plugins/GlobalEventsPlugin.tsx
