+++ src/script/components/RichTextEditor/plugins/HistoryPlugin.tsx
