+++ src/script/components/RichTextEditor/plugins/InlineEmojiReplacementPlugin/inlineReplacements.ts
