+++ src/script/components/RichTextEditor/plugins/MentionsPlugin/MentionSuggestionsItem.tsx
