+++ src/script/components/RichTextEditor/plugins/MentionsPlugin/MentionsPlugin.tsx
