+++ src/script/components/RichTextEditor/plugins/MentionsPlugin/index.ts
