+++ src/script/components/RichTextEditor/plugins/ReverseTypeaheadMenuPlugin.tsx
