+++ src/script/components/RichTextEditor/utils/generateNodes.ts
