+++ src/script/components/RichTextEditor/utils/getDomRangeRect.ts
