+++ src/script/components/RichTextEditor/utils/getSelectionInfo.ts
