+++ src/script/components/RichTextEditor/utils/getTextValue.ts
